Move RuntimeReward to compiler_gym.spaces
