+++ compiler_gym/spaces/__init__.py
@@ -9,6 +9,7 @@ from compiler_gym.spaces.discrete import Discrete
 from compiler_gym.spaces.named_discrete import NamedDiscrete
 from compiler_gym.spaces.permutation import Permutation
 from compiler_gym.spaces.reward import DefaultRewardFromObservation, Reward
+from compiler_gym.spaces.runtime_reward import RuntimeReward
 from compiler_gym.spaces.scalar import Scalar
 from compiler_gym.spaces.sequence import Sequence
 from compiler_gym.spaces.space_sequence import SpaceSequence
@@ -24,6 +25,7 @@ __all__ = [
     "NamedDiscrete",
     "Permutation",
     "Reward",
+    "RuntimeReward",
     "Scalar",
     "Sequence",
     "SpaceSequence",
